requirements.txt
accelerate==0.34.2
biopython==1.83
einops==0.8.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
transformers==4.37.2